Cleaned categories code
