--- client/src/components/categories/sections/AddCategorySection.tsx
@@ -1,9 +1,0 @@
-import { AddCategoryForm } from '../AddCategoryForm';
-
-export const AddCategoryItemsSection = () => {
-  return (
-    <>
-      <AddCategoryForm />
-    </>
-  );
-};
