--- client/src/components/categories/sections/CategoryTablesSection.tsx
@@ -1,9 +1,0 @@
-import { CategoryTables } from '../CategoryTables';
-
-export const CategoryTablesSection = () => {
-  return (
-    <>
-      <CategoryTables />
-    </>
-  );
-};
