+++ client/src/components/pages/layout/DefaultPageLayout.tsx
@@ -2,12 +2,10 @@ import { Box, Container } from '@mui/material';
 import { TopBar } from '../../top-bar/TopBar';
 import { Drawer } from '../../drawer/Drawer';
 import { Outlet } from 'react-router';
-import { AddCategoryModal } from '../../categories/AddCategoryModal';
 
 export const DefaultPageLayout = () => {
   return (
     <>
-      <AddCategoryModal />
       <TopBar />
       <Box display="flex">
         <Drawer />
